+++ app/client/src/utils/WidgetRegistry.tsx
@@ -89,7 +89,7 @@ import StatboxWidget, {
 } from "widgets/StatboxWidget";
 import FilePickerWidgetV2, {
   CONFIG as FILEPICKER_WIDGET_V2_CONFIG,
-} from "widgets/FilePickerWidgetV2Widget";
+} from "widgets/FilePickerWidgetV2";
 
 import AudioRecorderWidget, {
   CONFIG as AUDIO_RECORDER_WIDGET_CONFIG,
